moved the constraint code from ActionItem to RuleItem and added design elements for refinements
